Moved operator_str to core utils (not specific for ini)
